Updated Hour 10
respawns players. does not respawn balls. goals still register as being completed. this creates an error with the gui, since the game over script is constantly active.

need to make a cloning script for the balls, move the destroy command onto the balls, and figure out a way to synchronize the respawn across scripts.